+++ src/controller/ScoreControllerImpl.java
@@ -30,26 +30,23 @@ public final class ScoreControllerImpl implements ScoreController {
     private static final String SEPARATOR = File.separator;
 	private static final String FOLDER_NAME = System.getProperty("user.home") + SEPARATOR + "TLMS";
     private static final String FILE_NAME_RANKING = FOLDER_NAME + SEPARATOR + "ranking.json";
-    /**
-     * path of the user name file.
-     */
-    public static final String FILE_NAME_USER = FOLDER_NAME + SEPARATOR + "userName.json";
     private final MapController mapController = new MapControllerImpl();
 	private final ScoreModel scoreModel = new ScoreModelImpl();
+	private final UserNameController userNameController = new UserNameControllerImpl();
 	private List<String> list = new ArrayList<>();
 
 	@Override
 	public void firstGame() throws IOException {
 		try {
 			this.list = Files.lines(Paths.get(FILE_NAME_RANKING)).collect(Collectors.toList());
-			BufferedReader reader = new BufferedReader(new FileReader(FILE_NAME_USER));
+			BufferedReader reader = new BufferedReader(new FileReader(UserNameControllerImpl.FILE_NAME_USER));
 		    @SuppressWarnings("unused")
 		    final String userName = reader.readLine(); //unused but useful to know if it's the firt game ever
 		    reader.close();
 		} catch (IOException e) {
 			createFolder(FOLDER_NAME);
 			createRankingFile(FILE_NAME_RANKING);
-			createUserNameFile(FILE_NAME_USER);
+			createUserNameFile(UserNameControllerImpl.FILE_NAME_USER);
 		}	
 	}
 	
@@ -153,22 +150,5 @@ public final class ScoreControllerImpl implements ScoreController {
 				)
     	); 
     	return List.copyOf(scoreModel.rankingListRefactor(rankingList));
-	}
-
-	@Override
-	public String readUser() throws IOException {
-		try (BufferedReader reader = new BufferedReader(new FileReader(FILE_NAME_USER))) {
-			final String user = reader.readLine();
-			reader.close();
-			return user;
-		}
-	}
-
-	@Override
-	public void writeUser(final String user) throws IOException {
-		try (BufferedWriter writer = new BufferedWriter(new FileWriter(FILE_NAME_USER))) {
-			final String upperCaseUser = user.toUpperCase();
-			writer.write(upperCaseUser);
-		}
 	}
 }
